test(morph): cover chat transformation
